+++ modules/recheck-cli/src/test/scala/codes/quine/labo/recheck/cli/MainSuite.scala
@@ -3,7 +3,6 @@ package codes.quine.labo.recheck.cli
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.nio.charset.StandardCharsets
-import java.security.Permission
 import java.util.concurrent.TimeUnit
 import scala.concurrent.duration.Duration
 
@@ -35,40 +34,36 @@ class MainSuite extends munit.FunSuite {
     )
     assert(clue(Main.command.parse(Seq.empty)).isLeft)
     assertEquals(Main.command.parse(Seq("/foo/")), Right(Main.CheckAction(InputPattern("foo", ""), config)))
-    assertEquals(Main.command.parse(Seq("batch", "--thread-size=2")), Right(Main.BatchAction(2)))
-    assertEquals(Main.command.parse(Seq("batch")), Right(Main.BatchAction(sys.runtime.availableProcessors())))
+    assertEquals(Main.command.parse(Seq("agent", "--thread-size=2")), Right(Main.BatchAction(2)))
+    assertEquals(Main.command.parse(Seq("agent")), Right(Main.BatchAction(sys.runtime.availableProcessors())))
   }
 
-  test("Main.main") {
-    val oldSecurityManager = System.getSecurityManager
-    try {
-      class ExitException(val status: Int) extends SecurityException
-      System.setSecurityManager(new SecurityManager {
-        override def checkPermission(perm: Permission): Unit = {}
-        override def checkPermission(perm: Permission, context: Any): Unit = {}
-        override def checkExit(status: Int): Unit = throw new ExitException(status)
-      })
+  test("Main#run") {
+    class ExitException(val status: Int) extends SecurityException
 
-      def runMain(args: Seq[String], input: String = ""): (String, String, Int) = {
-        val newIn = new ByteArrayInputStream(input.getBytes(StandardCharsets.UTF_8))
-        val newOut = new ByteArrayOutputStream()
-        val newErr = new ByteArrayOutputStream()
+    def runMain(args: Seq[String], input: String = ""): (String, String, Int) = {
+      val newIn = new ByteArrayInputStream(input.getBytes(StandardCharsets.UTF_8))
+      val newOut = new ByteArrayOutputStream()
+      val newErr = new ByteArrayOutputStream()
 
-        val status =
-          try {
-            Console.withIn(newIn)(Console.withOut(newOut)(Console.withErr(newErr)(Main.main(args.toArray))))
-            0
-          } catch { case ex: ExitException => ex.status }
-        (newOut.toString(StandardCharsets.UTF_8), newErr.toString(StandardCharsets.UTF_8), status)
+      val main = new Main {
+        override def exit(exitCode: Int): Unit = throw new ExitException(exitCode)
       }
 
-      assertEquals(
-        runMain(Seq.empty),
-        ("", "Missing expected command (batch), or positional argument!\n\n" + Main.command.showHelp + "\n", 2)
-      )
-      assertEquals(runMain(Seq("/foo/"))._3, 0)
-      assertEquals(runMain(Seq("/(a|a)*$/"))._3, 1)
-      assertEquals(runMain(Seq("batch")), ("", "", 0))
-    } finally System.setSecurityManager(oldSecurityManager)
+      val status =
+        try {
+          Console.withIn(newIn)(Console.withOut(newOut)(Console.withErr(newErr)(main.run(args.toArray))))
+          0
+        } catch { case ex: ExitException => ex.status }
+      (newOut.toString(StandardCharsets.UTF_8), newErr.toString(StandardCharsets.UTF_8), status)
+    }
+
+    assertEquals(
+      runMain(Seq.empty),
+      ("", "Missing expected command (agent), or positional argument!\n\n" + Main.command.showHelp + "\n", 2)
+    )
+    assertEquals(runMain(Seq("/foo/"))._3, 0)
+    assertEquals(runMain(Seq("/(a|a)*$/"))._3, 1)
+    assertEquals(runMain(Seq("agent")), ("", "", 0))
   }
 }